Init Demo Commit
Hello Git-World!
Includes a small demo with minor assets. Demo allows you to move string
around the screen and fades a UI text gives a short control description
of the demo in and out. Due to simplicity and testing, no game objects
in this demo build take advantage of Delta Timer.

Some Asset files have been init in this commit.